network: honour RequiredForOnline= in network_is_online()
Ignore the operational state of interfaces which are not required for
online (cf. RequiredForOnline=) when checking if the system is online.
This is convenient for systemd-timesyncd - the only user of
network_is_online() - to only attempt a timesync when a suitable
interface has come online. This allows an administrator to set
RequiredForOnline=no for interfaces which do not provide any internet
connectivity, such as statically configured wireless access point
interfaces, and avoid spurious timesync attempts while other suitable
interfaces are still being configured.